Timeout fixes, round 2
So, despite some changes upstream to ddc, I was still experiencing
timeouts, and realized there was a bit of a flaw in my original thinking
for how to fix this issue.

The short of it - for every call to ddc#ale#get_completions, we actually
need to be sure we fire the notify callback before calling back into
Ale.  The reason for this - Ale will simply disregard previous
callbacks, and prep with the new ones.  While DDC did fix cases where
`gatherCandidates` will never return, it never accounts for the failure
of the `denops.call` failure.  Essentially the `once` callback MUST be
executed or else we'll see timeout errors for that registration.

I refactored both the VimL code and TypeScript code to work better
inside of this system.

For starters, in the VimL code, I hold onto the callback info locally,
and then if I get a new request for completions fire the callback
immediately with a `null` result.  I then updated the `once` handler to
accept this result and handle it properly, but resolving the promise
with an empty array.

As a side note, I realized even though I previously updated the ddc_vim
dependencies, there was actually even newer versions for us to use, so
did that cleanup as well.